+++ src/codegen-stubs.c
@@ -140,26 +140,24 @@ JL_DLLEXPORT void LLVMExtraAddDemoteFloat16Pass_fallback(void *PM) UNAVAILABLE
 JL_DLLEXPORT void LLVMExtraAddCPUFeaturesPass_fallback(void *PM) UNAVAILABLE
 
 //LLVM C api to the julia JIT
-JL_DLLEXPORT void* JLJITGetLLVMOrcExecutionSession_fallback(void* JIT) UNAVAILABLE
+JL_DLLEXPORT void* LLVMExtraJLJITGetLLVMOrcExecutionSession_fallback(void* JIT) UNAVAILABLE
 
-JL_DLLEXPORT void* JLJITGetJuliaOJIT_fallback(void) UNAVAILABLE
+JL_DLLEXPORT void* LLVMExtraJLJITGetJuliaOJIT_fallback(void) UNAVAILABLE
 
-JL_DLLEXPORT void* JLJITGetExternalJITDylib_fallback(void* JIT) UNAVAILABLE
+JL_DLLEXPORT void* LLVMExtraJLJITGetExternalJITDylib_fallback(void* JIT) UNAVAILABLE
 
-JL_DLLEXPORT void* JLJITAddObjectFile_fallback(void* JIT, void* JD, void* ObjBuffer) UNAVAILABLE
+JL_DLLEXPORT void* LLVMExtraJLJITAddObjectFile_fallback(void* JIT, void* JD, void* ObjBuffer) UNAVAILABLE
 
-JL_DLLEXPORT void* JLJITAddLLVMIRModule_fallback(void* JIT, void* JD, void* TSM) UNAVAILABLE
+JL_DLLEXPORT void* LLVMExtraJLJITAddLLVMIRModule_fallback(void* JIT, void* JD, void* TSM) UNAVAILABLE
 
-JL_DLLEXPORT void* JLJITLookup_fallback(void* JIT, void* Result, const char *Name) UNAVAILABLE
+JL_DLLEXPORT void* LLVMExtraJLJITLookup_fallback(void* JIT, void* Result, const char *Name) UNAVAILABLE
 
-JL_DLLEXPORT void* JLJITMangleAndIntern_fallback(void* JIT, const char *Name) UNAVAILABLE
+JL_DLLEXPORT void* LLVMExtraJLJITMangleAndIntern_fallback(void* JIT, const char *Name) UNAVAILABLE
 
-JL_DLLEXPORT const char *JLJITGetTripleString_fallback(void* JIT) UNAVAILABLE
+JL_DLLEXPORT const char *LLVMExtraJLJITGetTripleString_fallback(void* JIT) UNAVAILABLE
 
-JL_DLLEXPORT const char JLJITGetGlobalPrefix_fallback(void* JIT) UNAVAILABLE
+JL_DLLEXPORT const char LLVMExtraJLJITGetGlobalPrefix_fallback(void* JIT) UNAVAILABLE
 
-JL_DLLEXPORT const char *JLJITGetDataLayoutString_fallback(void* JIT) UNAVAILABLE
+JL_DLLEXPORT const char *LLVMExtraJLJITGetDataLayoutString_fallback(void* JIT) UNAVAILABLE
 
-JL_DLLEXPORT void* JLJITGetIRCompileLayer_fallback(void* JIT) UNAVAILABLE
-
-JL_DLLEXPORT void LLVMExtraOrcIRCompileLayerEmit_fallback(void* IRLayer, void* MR, void* TSM) UNAVAILABLE
+JL_DLLEXPORT void* LLVMExtraJLJITGetIRCompileLayer_fallback(void* JIT) UNAVAILABLE
